test: move backup tests into slow tests
Moves ITDatabaseAdminTests into the ITBackupTest which is only executed
on the nightly build
